Update the Meerkat BCT to support Nanya DDR chips
Rework the BCT to have only 2 configs, one for Micron DDR and one for
Nanya. The DDR type is selected by the strap pins on the Meerkat SOM.

Rename the BCT to remove the DDR chip name from the filename as we now
support 2 DDR types.

Finally remove config 0 and 1 as no SOM has been produced with these
strapping.

Change-Id: I1312dae43f2e335a967549fc2ad7bef4f2d09da2
Signed-off-by: Alban Bedel <alban.bedel@avionic-design.de>
Reviewed-on: http://review.adnet.avionic-design.de/4503
Reviewed-by: Julian Scheel <julian@jusst.de>